+++ Classes/GUI/NoHighlightImageAndTextCell.h
@@ -1,0 +1,18 @@
+//
+//  NoHighlightCell.h
+//  VGTagger
+//
+//  Created by Bilal Syed Hussain on 12/08/2011.
+//  Copyright 2011 St. Andrews KY16 9XW. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+#import "ImageAndTextCell.h"
+
+// An ImageAndTextCell that has no highlighting 
+@interface NoHighlightImageAndTextCell :ImageAndTextCell {
+@private
+    
+}
+
+@end