+++ Classes/GUI/NoHighlightImageAndTextCell.m
@@ -1,0 +1,20 @@
+//
+//  NoHighlightCell.m
+//  VGTagger
+//
+//  Created by Bilal Syed Hussain on 12/08/2011.
+//  Copyright 2011 St. Andrews KY16 9XW. All rights reserved.
+//
+
+#import "NoHighlightImageAndTextCell.h"
+
+
+@implementation NoHighlightImageAndTextCell
+
+
+- (NSColor *)highlightColorWithFrame:(NSRect)cellFrame inView:(NSView *)controlView
+{
+	return nil;
+}
+
+@end