+++ Classes/GUI/QuickLookTableView.m
@@ -10,6 +10,7 @@
 #import "VGTaggerAppDelegate.h"
 #import "MainController.h"
 #import "FileSystemNode.h"
+#import "BHFinderLabelColours.h"
 
 @implementation QuickLookTableView
 
@@ -23,51 +24,85 @@
     }
 }
 
+- (void)highlightSelectionInClipRect:(NSRect)clipRect
+{
+	NSIndexSet *selectedRowIndexes = [self selectedRowIndexes];
+	NSRange visibleRows = [self rowsInRect:clipRect];
+	
+	NSUInteger selectedRow = [selectedRowIndexes firstIndex];
+    while (selectedRow != NSNotFound)
+    {
+		if (selectedRow == -1 || !NSLocationInRange(selectedRow, visibleRows)) {
+			selectedRow = [selectedRowIndexes indexGreaterThanIndex:selectedRow];
+			continue;
+		}		
+		
+		if ( [[self window] isKeyWindow] ) {
+			[[NSColor alternateSelectedControlColor] set];
+		}
+		else {
+			[[NSColor secondarySelectedControlColor] set];
+		}
+		
+		NSRect rowRect = [self rectOfRow:selectedRow];
+		NSRectFill(rowRect);
+		
+		// draw a "dot" of color, similar to what the Finder does, 
+		NSInteger labelIndex = BHLabelColorNone;
+		if ([[self dataSource] respondsToSelector:@selector(labelColorForRow:)]) {
+			labelIndex = [(id)[self dataSource] labelColorForRow:selectedRow];
+		}
+		NSGradient *gradient = [BHFinderLabelColours gradientForLabelIndex:labelIndex];
+		
+		if (gradient)
+		{			
+			NSRect gradientRect = 
+			NSMakeRect(2, 
+					   ([self rowHeight]+[self intercellSpacing].height)*selectedRow + 1.0, 
+					   20.0, 
+					   rowRect.size.height - 2.0);
+			NSBezierPath *gradientPath = [NSBezierPath bezierPathWithRoundedRect:gradientRect 
+																		 xRadius:9.0 
+																		 yRadius:9.0];
+			[gradient drawInBezierPath:gradientPath angle:90.0];			
+		}
+		
+		selectedRow = [selectedRowIndexes indexGreaterThanIndex:selectedRow];
+    }
+}
+
 // Override The row colour with the label colour for only the filename column.
 // with nice rounded side if there is a label colour.
 - (void)drawRow:(NSInteger)row clipRect:(NSRect)clipRect;
 {
-	NSColor *colour = [[[[[(MainController*)[self dataSource] directoryStack] 
-					   lastObject] children] 
-	  objectAtIndex:row] labelColor];
-	if (!colour){
-		[super drawRow:row clipRect:clipRect];
-		return;
-	};
-	
-	[NSGraphicsContext saveGraphicsState];
-	NSRectClip(clipRect);
-	// draw over the alternating row colour.
-	[[NSColor whiteColor] setFill];
-	
-	NSRect rowRect = [self rectOfRow:row];
-	// colour only the filename column.
-	NSRect colRect = [self rectOfColumn: [self columnWithIdentifier:@"filename"]];
-	rowRect.origin.x   = colRect.origin.x;
-	
-	// Only cover the background of the image.
-	NSBezierPath *p;
-	if (row ==[self selectedRow]){
-		rowRect.size.width = 21;
-		NSRectFill(NSIntersectionRect(rowRect, clipRect));
-		p  = [NSBezierPath bezierPathWithRect:rowRect];
-	}else{
-		rowRect.size.width = colRect.size.width;
-		NSRectFill(NSIntersectionRect(rowRect, clipRect));
-		// draw with rounded end caps
-		CGFloat radius = NSHeight(rowRect) / 2.0;
-		p = [NSBezierPath bezierPathWithRoundedRect:NSInsetRect(rowRect, 1.0, 0.0) 
-														  xRadius:radius 
-														  yRadius:radius];
+	NSInteger labelIndex= BHLabelColorNone;
+	if ([[self dataSource] respondsToSelector:@selector(labelColorForRow:)]) {
+		labelIndex = [(id)[self dataSource] labelColorForRow:row];
 	}
 	
-	[colour setFill];
-	[p fill];
-	[NSGraphicsContext restoreGraphicsState];
+	NSGradient *gradient = [BHFinderLabelColours gradientForLabelIndex:labelIndex];
 	
-	// draw cells on top of the new row background
+	if (gradient && ![[self selectedRowIndexes] containsIndex:row])
+	{
+		NSRect rowRect = [self rectOfRow:row];
+		NSRect colRect = [self rectOfColumn: [self columnWithIdentifier:@"filename"]];
+		rowRect.origin.x   = colRect.origin.x;
+		rowRect.size.width = colRect.size.width;
+
+		NSRect gradientRect = NSMakeRect(0, 
+										 ([self rowHeight]+[self intercellSpacing].height)*row + 1.0, 
+										 rowRect.size.width - 4.0, 
+										 rowRect.size.height - 2.0);
+		NSBezierPath *smallerPath = [NSBezierPath bezierPathWithRoundedRect:gradientRect 
+																	xRadius:9.0 
+																	yRadius:9.0];
+		[gradient drawInBezierPath:smallerPath 
+							 angle:90.0];
+	}
+	
 	[super drawRow:row clipRect:clipRect];
 }
+
 
 
 @end