+++ Changelog.md
@@ -4,10 +4,14 @@ Changelog
 -----------
 Version 1.3
 -----------
+### Features ###
 * Label colours for files shown.
 * Nicer display of label colour for selected row.
 * Right click menu that allows labeling files.
 * Right click menu item reveal in finder.
+
+### Improvements ###
+* Shows the label colour as a dot when selecting a labeled row.
 
 -----------
 Version 1.2 (Thu Aug 11 2011 03:34:20 +0100)